Implementing new Order Card widget
